requirements.txt
flask==3.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
faiss-cpu==1.7.4
sentence-transformers==2.5.1
pandas==2.1.4
numpy==1.26.3
spacy==3.7.2
PyPDF2==3.0.1
pt-core-news-md @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_md-3.7.0/pt_core_news_md-3.7.0.tar.gz